+++ packages/illustrations/CHANGELOG.md
@@ -8,6 +8,20 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.9.0 (6/24/2024 PST)
+
+#### 🚀 Updates
+
+- Publish illustrations 06/24/2024. [[#3109](https://github.cbhq.net/frontend/cds/pull/3109)]
+
+### 🐞 Updated
+
+Generated with `yarn nx run illustrations:sync`
+
+#### Spot Rectangle
+
+- appUpdate
+
 ## 3.8.0 (5/31/2024 PST)
 
 #### 🚀 Updates